Breaking pairs out into its own module
