+++ include/lispint.h
@@ -3,7 +3,7 @@
 #ifndef FILE_LISPINT_SEEN
 #define FILE_LISPINT_SEEN
 
-object *make_fixnum(long value);
-bool is_fixnum(object *obj);
+extern object *make_fixnum(long value);
+extern bool is_fixnum(object *obj);
 
 #endif // !FILE_LISPINT_SEEN
